Add files via upload
This file contains the code for A * search algorithm. 
A* Search algorithm is one of the best and popular technique used in path-finding and graph traversals. 
The A* algorithm is a search algorithm that finds the shortest path between two points on a graph. It's often used for map traversal and path planning.